Try dartfmt

+++ packages/flutter_tools/lib/src/bundle.dart
@@ -21,9 +21,10 @@ import 'project.dart';
 const String defaultMainPath = 'lib/main.dart';
 const String defaultAssetBasePath = '.';
 const String defaultManifestPath = 'pubspec.yaml';
-String get defaultDepfilePath => fs.path.join(getBuildDirectory(), 'snapshot_blob.bin.d');
+String get defaultDepfilePath =>
+    fs.path.join(getBuildDirectory(), 'snapshot_blob.bin.d');
 
-String getDefaultApplicationKernelPath({ @required bool trackWidgetCreation }) {
+String getDefaultApplicationKernelPath({@required bool trackWidgetCreation}) {
   return getKernelPathForTransformerOptions(
     fs.path.join(getBuildDirectory(), 'app.dill'),
     trackWidgetCreation: trackWidgetCreation,
@@ -69,27 +70,29 @@ Future<void> build({
   depfilePath ??= defaultDepfilePath;
   assetDirPath ??= getAssetBuildDirectory();
   packagesPath ??= fs.path.absolute(PackageMap.globalPackagesPath);
-  applicationKernelFilePath ??= getDefaultApplicationKernelPath(trackWidgetCreation: trackWidgetCreation);
+  applicationKernelFilePath ??=
+      getDefaultApplicationKernelPath(trackWidgetCreation: trackWidgetCreation);
   final FlutterProject flutterProject = FlutterProject.current();
 
   if (compilationTraceFilePath != null) {
-    if (buildMode != BuildMode.dynamicProfile && buildMode != BuildMode.dynamicRelease) {
+    if (buildMode != BuildMode.dynamicProfile &&
+        buildMode != BuildMode.dynamicRelease) {
       // Silently ignore JIT snapshotting for those builds that don't support it.
       compilationTraceFilePath = null;
-
     } else if (compilationTraceFilePath.isEmpty) {
       // Disable JIT snapshotting if flag is empty.
       printStatus('Code snapshot will be disabled for this build.');
       compilationTraceFilePath = null;
-
     } else if (!fs.file(compilationTraceFilePath).existsSync()) {
       // Be forgiving if compilation trace file is missing.
-      printStatus('No compilation trace available. To optimize performance, consider using --train.');
-      final File tmp = fs.systemTempDirectory.childFile('flutterEmptyCompilationTrace.txt');
+      printStatus(
+          'No compilation trace available. To optimize performance, consider using --train.');
+      final File tmp =
+          fs.systemTempDirectory.childFile('flutterEmptyCompilationTrace.txt');
       compilationTraceFilePath = (tmp..createSync(recursive: true)).path;
-
     } else {
-      printStatus('Code snapshot will use compilation training file $compilationTraceFilePath.');
+      printStatus(
+          'Code snapshot will use compilation training file $compilationTraceFilePath.');
     }
   }
 
@@ -98,11 +101,14 @@ Future<void> build({
     if ((extraFrontEndOptions != null) && extraFrontEndOptions.isNotEmpty)
       printTrace('Extra front-end options: $extraFrontEndOptions');
     ensureDirectoryExists(applicationKernelFilePath);
-    final KernelCompiler kernelCompiler = await kernelCompilerFactory.create(flutterProject);
+    final KernelCompiler kernelCompiler =
+        await kernelCompilerFactory.create(flutterProject);
     final CompilerOutput compilerOutput = await kernelCompiler.compile(
-      sdkRoot: artifacts.getArtifactPath(Artifact.flutterPatchedSdkPath, mode: buildMode),
-      incrementalCompilerByteStorePath: compilationTraceFilePath != null ? null :
-          fs.path.absolute(getIncrementalCompilerByteStoreDirectory()),
+      sdkRoot: artifacts.getArtifactPath(Artifact.flutterPatchedSdkPath,
+          mode: buildMode),
+      incrementalCompilerByteStorePath: compilationTraceFilePath != null
+          ? null
+          : fs.path.absolute(getIncrementalCompilerByteStoreDirectory()),
       mainPath: fs.file(mainPath).absolute.path,
       outputFilePath: applicationKernelFilePath,
       depFilePath: depfilePath,
@@ -118,8 +124,11 @@ Future<void> build({
     }
     kernelContent = DevFSFileContent(fs.file(compilerOutput.outputFilename));
 
-    await fs.directory(getBuildDirectory()).childFile('frontend_server.d')
-        .writeAsString('frontend_server.d: ${artifacts.getArtifactPath(Artifact.frontendServerSnapshotForEngineDartSdk)}\n');
+    await fs
+        .directory(getBuildDirectory())
+        .childFile('frontend_server.d')
+        .writeAsString(
+            'frontend_server.d: ${artifacts.getArtifactPath(Artifact.frontendServerSnapshotForEngineDartSdk)}\n');
 
     if (compilationTraceFilePath != null) {
       final JITSnapshotter snapshotter = JITSnapshotter();
@@ -133,7 +142,8 @@ Future<void> build({
         extraGenSnapshotOptions: extraGenSnapshotOptions,
       );
       if (snapshotExitCode != 0) {
-        throwToolExit('Snapshotting exited with non-zero exit code: $snapshotExitCode');
+        throwToolExit(
+            'Snapshotting exited with non-zero exit code: $snapshotExitCode');
       }
     }
   }
@@ -144,8 +154,7 @@ Future<void> build({
     packagesPath: packagesPath,
     reportLicensedPackages: reportLicensedPackages,
   );
-  if (assets == null)
-    throwToolExit('Error building assets', exitCode: 1);
+  if (assets == null) throwToolExit('Error building assets', exitCode: 1);
 
   await assemble(
     buildMode: buildMode,
@@ -176,8 +185,7 @@ Future<AssetBundle> buildAssets({
     includeDefaultFonts: includeDefaultFonts,
     reportLicensedPackages: reportLicensedPackages,
   );
-  if (result != 0)
-    return null;
+  if (result != 0) return null;
 
   return assetBundle;
 }
@@ -193,21 +201,32 @@ Future<void> assemble({
   assetDirPath ??= getAssetBuildDirectory();
   printTrace('Building bundle');
 
-  final Map<String, DevFSContent> assetEntries = Map<String, DevFSContent>.from(assetBundle.entries);
+  final Map<String, DevFSContent> assetEntries =
+      Map<String, DevFSContent>.from(assetBundle.entries);
   if (kernelContent != null) {
     if (compilationTraceFilePath != null) {
-      final String vmSnapshotData = artifacts.getArtifactPath(Artifact.vmSnapshotData, mode: buildMode);
-      final String isolateSnapshotData = fs.path.join(getBuildDirectory(), _kIsolateSnapshotData);
-      final String isolateSnapshotInstr = fs.path.join(getBuildDirectory(), _kIsolateSnapshotInstr);
-      assetEntries[_kVMSnapshotData] = DevFSFileContent(fs.file(vmSnapshotData));
-      assetEntries[_kIsolateSnapshotData] = DevFSFileContent(fs.file(isolateSnapshotData));
-      assetEntries[_kIsolateSnapshotInstr] = DevFSFileContent(fs.file(isolateSnapshotInstr));
+      final String vmSnapshotData =
+          artifacts.getArtifactPath(Artifact.vmSnapshotData, mode: buildMode);
+      final String isolateSnapshotData =
+          fs.path.join(getBuildDirectory(), _kIsolateSnapshotData);
+      final String isolateSnapshotInstr =
+          fs.path.join(getBuildDirectory(), _kIsolateSnapshotInstr);
+      assetEntries[_kVMSnapshotData] =
+          DevFSFileContent(fs.file(vmSnapshotData));
+      assetEntries[_kIsolateSnapshotData] =
+          DevFSFileContent(fs.file(isolateSnapshotData));
+      assetEntries[_kIsolateSnapshotInstr] =
+          DevFSFileContent(fs.file(isolateSnapshotInstr));
     } else {
-      final String vmSnapshotData = artifacts.getArtifactPath(Artifact.vmSnapshotData, mode: buildMode);
-      final String isolateSnapshotData = artifacts.getArtifactPath(Artifact.isolateSnapshotData, mode: buildMode);
+      final String vmSnapshotData =
+          artifacts.getArtifactPath(Artifact.vmSnapshotData, mode: buildMode);
+      final String isolateSnapshotData = artifacts
+          .getArtifactPath(Artifact.isolateSnapshotData, mode: buildMode);
       assetEntries[_kKernelKey] = kernelContent;
-      assetEntries[_kVMSnapshotData] = DevFSFileContent(fs.file(vmSnapshotData));
-      assetEntries[_kIsolateSnapshotData] = DevFSFileContent(fs.file(isolateSnapshotData));
+      assetEntries[_kVMSnapshotData] =
+          DevFSFileContent(fs.file(vmSnapshotData));
+      assetEntries[_kIsolateSnapshotData] =
+          DevFSFileContent(fs.file(isolateSnapshotData));
     }
   }
 
@@ -222,14 +241,13 @@ Future<void> writeBundle(
   Directory bundleDir,
   Map<String, DevFSContent> assetEntries,
 ) async {
-  if (bundleDir.existsSync())
-    bundleDir.deleteSync(recursive: true);
+  if (bundleDir.existsSync()) bundleDir.deleteSync(recursive: true);
   bundleDir.createSync(recursive: true);
 
-  await Future.wait<void>(
-    assetEntries.entries.map<Future<void>>((MapEntry<String, DevFSContent> entry) async {
-      final File file = fs.file(fs.path.join(bundleDir.path, entry.key));
-      file.parent.createSync(recursive: true);
-      await file.writeAsBytes(await entry.value.contentsAsBytes());
-    }));
+  await Future.wait<void>(assetEntries.entries
+      .map<Future<void>>((MapEntry<String, DevFSContent> entry) async {
+    final File file = fs.file(fs.path.join(bundleDir.path, entry.key));
+    file.parent.createSync(recursive: true);
+    await file.writeAsBytes(await entry.value.contentsAsBytes());
+  }));
 }
